agents/frontend-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.25.2
pydantic==2.5.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.2
